Moved cursor.posX label and fixed text label. Fixed value and style of statusbar
Moved the cursor.posX label to behind the "lines" label.
Changed the text to the correct label to "Col" to reflect how all other editors do it.
Fixed the starting value of posX to 1 instead of 0
Restyled the statusbar so that labels are italic and values are bold.
Making it easier to see the important info.